Add .contain class
We removed overflow: hidden from <article>. So sometimes we'll
need to prevent items inside <article> from leaking when they grow
to be too big. When they leak, they can cause the viewport to slide
on mobile. So we can wrap those items in a <div class="contain"> to
prevent this.